pyproject.toml
[project]
name = "cipherswarm"
version = "0.1.0"
description = "A distributed password cracking management system"
authors = [{ name = "CipherSwarm Team", email = "team@cipherswarm.org" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.115.12",
    "uvicorn[standard]>=0.27.1",
    "sqlalchemy>=2.0.27",
    "alembic>=1.15.2",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "celery>=5.3.6",
    "redis>=5.0.1",
    "httpx[http2,brotli,zstd]~=0.28.1",
    "python-dotenv>=1.0.1",
    "email-validator>=2.1.0.post1",
    "psycopg2-binary>=2.9.10",
    "loguru>=0.7.3",
    "casbin>=1.41.0",
    "pyjwt>=2.10.1",
    "bcrypt==4.0.1",
    "cashews>=7.4.0",
    "packaging>=25.0",
    "timeago>=1.0.16",
]

[dependency-groups]
dev = [
    "mkdocs-exporter>=6.2.0",
    "mkdocstrings[python]>=0.29.1",
    "pre-commit>=4.2.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.5",
    "pytest-postgresql>=5.0.0",
    "black>=24.1.1",
    "ruff>=0.2.1",
    "mypy>=1.8.0",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.5.13",
    "polyfactory>=2.21.0",
    "types-passlib>=1.7.7.20250408",
    "types-python-jose>=3.4.0.20250224",
]

[project.scripts]
server = "app.main:run_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py313"

exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    "alembic/versions/*",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = [
    "ALL", # include all the rules, including new ones
]
ignore = [
    "B008",    # doesn't play well with FastAPI DI
    "COM812",  # Runs counter to Ruff formatting rules
    "D100",    # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D203",
    "D205",
    "D212",
    "D400",
    "D401",
    "D415",
    "E402",    # false positives for local imports
    "E501",    # line too long
    "EM102",   # Stupid error that doesn't make sense
    "EM101",   # Stupid error that doesn't make sense
    "FBT001",  # Stupid rule that doesn't make sense
    "FBT002",  # Stupid rule that doesn't make sense
    "FBT003",  # Stupid rule that doesn't make sense
    "G004",    # I don't care about this rule
    "TRY003",  # external messages in exceptions are too verbose
    "TC001",
    "TC002",   # too many false positives from my own code
    "TC003",
    "TD002",
    "TD003",
    "TRY301",  # Not sure why this is an error
    "PLR0913", # too many arguments is sometimes fine
    "PLR0915", # too many arguments is sometimes fine
    "FIX002",  # Temporary ignore for ruff while I'm still implementing the changes
    "ERA001",  # Temporary ignore for ruff while I'm still implementing the changes
]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "S", "PT011", "ARG001", "SLF001", "ANN401", "T201"]
"alembic/versions/**/*.py" = ["D"]
"tests/conftest.py" = ["ARG001", "SLF001", "ANN401"]
"tests/db/test_session.py" = ["SLF001"]
"scripts/dev/**/*.py" = ["T201"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
exclude = '^(\\.venv|build|dist|alembic)/'
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disable_error_code = [
    "attr-defined",
    "no-untyped-def",
    "var-annotated",
    "call-arg",
    "return",
]

[[tool.mypy.overrides]]
module = "tests.factories.*"
disable_error_code = ["unused-ignore", "syntax"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=app"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
